Search PR review refactoring proposal
