--- components/search/context.ts
@@ -1,8 +1,0 @@
-import { type Dispatch, type SetStateAction, createContext } from 'react';
-
-const SearchFocusedContext = createContext<{
-  isFocused: boolean;
-  setIsFocused: Dispatch<SetStateAction<boolean>>;
-}>({ isFocused: false, setIsFocused: () => {} });
-
-export default SearchFocusedContext;
