+++ components/search/widgets/hit/models.ts
