+++ components/search/widgets/hit/sectionIcon.tsx
