+++ components/search/widgets/hit/useHit.ts
@@ -1,26 +1,30 @@
 import { useRouter } from 'next/router';
-import { useCallback, useContext, useEffect, useState } from 'react';
+import { useCallback, useEffect, useState } from 'react';
 
-import SearchFocusedContext from './context';
+import useFocus from '../../useFocus';
 
 export default function useHit() {
   const router = useRouter();
-  const { setIsFocused } = useContext(SearchFocusedContext);
+  const { setIsFocused } = useFocus();
   const [isLoading, setIsLoading] = useState(false);
+
   const handleLoadingEnded = useCallback(() => {
     setIsLoading(false);
     router.events.off('routeChangeComplete', handleLoadingEnded);
     setIsFocused(false);
   }, [router.events, setIsFocused]);
+
   const handleClick = useCallback(() => {
     setIsLoading(true);
     router.events.on('routeChangeComplete', handleLoadingEnded);
   }, [router.events, handleLoadingEnded]);
+
   useEffect(
     () => () => {
       router.events.off('routeChangeComplete', handleLoadingEnded);
     },
     [router.events, handleLoadingEnded]
   );
+
   return { handleClick, isLoading };
 }
