+++ components/search/widgets/noResultsBoundary.tsx
